Large refactor around require block formatting
This was a large problem and resulted in being very brittle to use with
the formatting being wrong more often than not. This refactor aims to
fix this by constructing the white space intelligently and replacing the
entire contents from first to second bracket.

Also, if no require paths have been loaded, to give a strong starting
point we now replace the white space with a single new line.